Fix casting bugs for XCode6b2

+++ BNRHypnosister-Swift/BNRHypnosisView.swift
@@ -29,9 +29,9 @@ class BNRHypnosisView: UIView {
         println("\(self) was touched")
         
         //Get 3 random numbers between 0 and 1
-        let red = Double(arc4random() % 100) / 100.0
-        let green = Double(arc4random() % 100) / 100.0
-        let blue = Double(arc4random() % 100) / 100.0
+        let red = CGFloat(arc4random() % 100) / 100.0
+        let green = CGFloat(arc4random() % 100) / 100.0
+        let blue = CGFloat(arc4random() % 100) / 100.0
         
         let randomColor = UIColor(red: red, green: green, blue: blue, alpha: 1.0)
         
@@ -51,14 +51,14 @@ class BNRHypnosisView: UIView {
         center.y = bounds.origin.y + bounds.size.height / 2.0
         
         //The largest circle will circumscribe the view
-        let maxRadius : Double  = hypot(bounds.size.width, bounds.size.height) / 2.0
+        let maxRadius : CGFloat  = CGFloat(hypot(Double(bounds.size.width), Double(bounds.size.height))) / 2.0
     
 
         let path = UIBezierPath()
         
         for var currentRadius = maxRadius; currentRadius > 0; currentRadius -= 20 {
             path.moveToPoint(CGPointMake(center.x + currentRadius, center.y))   
-            path.addArcWithCenter(center, radius: currentRadius, startAngle: 0.0, endAngle: M_PI * 2.0, clockwise: true)
+            path.addArcWithCenter(center, radius: currentRadius, startAngle: 0.0, endAngle: CGFloat(M_PI * 2.0), clockwise: true)
             //Note this is currentRadius!
         }
         //Configure line width ot 10 points
